Create stories
Four tales {Klaas(0), Cinderella(1), Musicians (2), Give me(3)}, each told first in a Cause version (0) then in an Intention (1) version. (Note: 'a' stands for 'aspect', a term we chose meaning facet or slant, before we realised it also has a specific linguistics meaning. This is not connected to the linguistics meaning).